Move bind functions to separate .h file
